backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.0
Pygments==2.17.2
sqlglot==20.4.0
gunicorn==20.1.0